--- draw.cpp
@@ -1,2 +1,0 @@
-#include "draw.h"
-
